Finish step 2
Translating the Vec3f code from C++ to Rust hasn't been completely
straightforward. C++ seems to have many more overloaded methods built in, making
it easy to do math on them and cast them to floats out of the box.

I've found it useful to look at a raytracer built in Rust[1] to help out with the
syntax differences.

[1]:https://github.com/bheisler/raytracer